+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changes
 2.0.0 (unreleased)
 ------------------
 
-- Use taxonomy default language if current or target language is not available
+- Use taxonomy default language for indexing if current language is not available
   [agitator]
 
 - Make generated behaviors language independent