+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changes
 =======
 
-2.0.0 (unreleased)
+2.0.0 (2019-11-25)
 ------------------
 
 - add uninstall handler to cleanup persistent utilites and generated behaviors